+++ examples/file-transfer/sender.rs
@@ -1,7 +1,7 @@
 use super::common::{SenderMsg, ReceiverMsg};
 
 use message_io::events::{EventQueue};
-use message_io::network::{Network, NetEvent};
+use message_io::network::{Network, NetEvent, Transport};
 
 use std::fs::{self, File};
 use std::io::{Read};
@@ -18,7 +18,7 @@ pub fn run(file_path: &str) {
     let mut network = Network::new(move |net_event| network_sender.send(Event::Network(net_event)));
 
     let server_addr = "127.0.0.1:3005";
-    let server_id = match network.connect_tcp(server_addr) {
+    let server_id = match network.connect(Transport::Tcp, server_addr) {
         Ok(server_id) => {
             println!("Connect to receiver by TCP at {}", server_addr);
             server_id